Fixed Windows default spinner and prefix (#8)
* Fixed Windows default spinner and prefix

On Windows the default spinner was never correctly set to "dashes", and was always set to "dots". This has been fixed. Also the default command line cannot handle the current characters used for success and fail prefixes, these have been updated to check the system type and allows you to set custom prefixes

* Fixed prefixLength and removed unused deconstruction