:white_check_mark: Add tests
Write tests that test spinner modify functions with spinners.get('name').

Fix README.md indent.

Fix statusOptionsFromNormalUpdate, it now sets the textColor option for fail and succeed statuses.

statusOptionsFromNormalUpdate will run before the spinner options were purged, in spinner.update(). Which is why it will also check types.
It fixes the issue where failPrefix and succeedPrefix options will be purged and not apply for the statusOverride.

Fix the tests for statusOptionsFromNormalUpdate, to match the changes to statusOptionsFromNormalUpdate.