:strawberry: spinnie.rawRender()
Create the rawRender method on the spinnie class. It will try to get a rawRender function from the current status and default to rendering the prefix with the text.

rawRender on the status options should take care of the coloring. While everything else like indent and breaking text will be applied by the spinnie.rawRender method.

Status option overrides will now apply on creation and not only when a spinner is updated.

update statusOptionsFromNormalUpdate to update based on the spinnerColor option.

prefixColor and spinnerColor will default to cyan instead of greenBright.

Note: setRawStreamOutput is not the final form of CI/not-spinning rendering.