:wolf: Implement the use of spinner.statusName()
Hooray!

So how is this done?
StatusRegistry now extends EventEmitter and emits an event name 'statusAdded' whenever you add a status.
The main spinnies instance will listen to that event and loop the spinners and use aliasStatusAsMethod to alias that status. It will also use aliasChildMethod to alias spinnies.statusName('spinnerName') to spinnies.get('spinnerName').statusName().

Each spinner will use aliasStatusAsMethod on initialization for each of the statuses (and aliases) in the statusRegistry.

TODO: write tests