:b: Create the 'bind' method
Pass a promise to 'bind'.

If that promise resolves it will succeed the spinner with an optional success message.

If that promise rejects it will fail the spinner with an optional error message.

This is a very basic implementation of the bind method.
It should be able to handle an observable and to format error messages if an error is passed.